refactor(constants): isolate each provider definition into its own file
