--- packages/constants/src/providers/apiKeys.ts
@@ -1,8 +1,0 @@
-export const NEOSANTARA_BASE_URL = "https://api.neosantara.xyz/v1";
-export const GOROUTER_BASE_URL = "https://gorouter.app/v1";
-export const BLUESMINDS_BASE_URL = "https://api.bluesminds.com/v1";
-export const SEEKAI_BASE_URL = "https://seekai.cc/v1";
-export const TABITOKEN_BASE_URL = "https://tabitoken.com/v1";
-export const TOKENROUTER_BASE_URL = "https://api.tokenrouter.com/v1";
-export const COMMANDCODE_BASE_URL = "https://api.commandcode.ai/alpha/generate";
-export const COMMANDCODE_MODELS_URL = "https://api.commandcode.ai/provider/v1/models";
